requirements.txt
pandas==1.2.5
scikit-learn==1.2.2
Pillow==9.5.0
matplotlib==3.7.1
nbconvert==7.4.0
seaborn==0.12.2
